refactor(weather): Move screens under view/screens directory for better separation
- Moved city_search_screen.dart, error_display_screen.dart, and weather_result_screen.dart
to features/weather/view/screens/ to align with clean architecture conventions.
- Updated relative import paths accordingly.
